% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmds.diss.R
\name{cmds.diss}
\alias{cmds.diss}
\title{Classical Scaling of Dissimilarity Measures}
\usage{
cmds.diss(
  data,
  k = ncol(data),
  x.use = FALSE,
  zero.chk = TRUE,
  plt = FALSE,
  plot.subset = FALSE,
  plot.subn = 5,
  ...
)
}
\arguments{
\item{data}{Data matrix.}

\item{k}{Number of vectors to be returned.}

\item{x.use}{Whether to use extended dissimilarity.}

\item{zero.chk}{Check for zero row sums -- if zero ignore these rows
according to method.}

\item{plt}{Whether to plot the relationship between the dissimilarities and
the distances calculated from the scaled output vectors.}

\item{plot.subset}{Plot a subset of the points -- useful for large data sets.}

\item{plot.subn}{Controls how many points are plotted when
\code{plot.subset = TRUE}. The number of points plotted is
\code{750 + N * plot.subn} where N = number of rows in \code{data}.}

\item{...}{Arguments passed to either \code{xdiss} or \code{gdist}.}
}
\description{
The function first computes the dissimilarity matrix according to the
specified method -- see \code{mvpart}. The dissimilarities are then scaled
using classical scaling -- see \code{\link[stats]{cmdscale}}. The returned
matrix can be input into \code{mvpart} for multivariate regression tree
splitting.
}
\details{
The function knows the same dissimilarity indices as
\code{mvpart}. Plotting the relationship between the dissimilarities and the
distances calculated from the scaled output vectors is useful in assessing
potential loss of information. If the loss is high then the results from
partitioning directly from the dissimilarity matrix using distance-base
partitioning (see \code{dist} in \code{mvpart}), and those obtained from
partitioning the output of \code{cmds.diss} using multivariate regression
trees (see \code{mrt} in \code{mvpart}) can be substantial.
}
\examples{
data(spider)
dist.vecs <- cmds.diss(spider)

}
\author{
Glenn De'ath
}
\keyword{multivariate}
