% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.rpart.R
\name{na.rpart}
\alias{na.rpart}
\title{Handles Missing Values in an Rpart Object}
\usage{
na.rpart(x)
}
\arguments{
\item{x}{A model frame.}
}
\description{
Handles missing values in an \code{rpart} object.
}
\details{
Internal function that handles missing values when calling the
function \code{rpart}.
}
\keyword{tree}
