% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{rpart-generics}
\alias{rpart-generics}
\alias{prune}
\alias{rsq}
\alias{meanvar}
\title{Generic Methods}
\usage{
prune(object, ...)

rsq(x, ...)

meanvar(object, ...)
}
\arguments{
\item{object}{An object of class \code{rpart}. This is assumed to be the
result of some function that produces an object with the same named
components as that returned by the \code{rpart} function.}

\item{...}{Optional arguments to be passed internally to other functions.}

\item{x}{Identical as argument \code{object}.}
}
\value{
\describe{
  \item{prune}{...}
  \item{rsq}{...}
  \item{meanvar}{...}
}
}
\description{
A set of generic functions used by mvpart.
}
\section{Functions}{
\itemize{
\item \code{prune()}: Prune Method

Method to prune branch for a regression or classification tree

\item \code{rsq()}: R-square Method

Method to show the variation of the R-square with the number of splits.

\item \code{meanvar()}: Meanvar method

[Description of the meanvar generic here.]

}}
\seealso{
\code{\link{prune.rpart}} \code{\link{rsq.rpart}}, and
\code{meanvar.rpart}
}
