% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.matrix.R
\name{rpart.matrix}
\alias{rpart.matrix}
\title{Creates a Model Matrix from a Call to Rpart.}
\usage{
rpart.matrix(frame)
}
\arguments{
\item{frame}{model frame (from call to \code{rpart}).}
}
\value{
Returns a matrix from the frame of class \code{matrix}.  Used
internally by the function \code{rpart}.
}
\description{
Creates a model matrix from a call to \code{rpart}.
}
\keyword{tree}
