% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart-methods.R
\name{rpart.methods}
\alias{rpart.methods}
\alias{rpart.anova}
\alias{rpart.mrt}
\alias{rpart.poisson}
\alias{rpart.class}
\alias{rpart.dist}
\alias{rpart.exp}
\title{Set up Methods for an rpart Model}
\usage{
rpart.anova(y, offset, parms, wt)

rpart.mrt(y, offset, parms, wt)

rpart.poisson(y, offset, parms, wt)

rpart.class(y, offset, parms, wt)

rpart.dist(y, offset, parms, wt)

rpart.exp(y, offset, parms, wt)
}
\arguments{
\item{y}{The responses.}

\item{offset}{An offset, or \code{NULL}.}

\item{parms}{A list of parameters, usually empty.}

\item{wt}{Case weights.}
}
\value{
\describe{
  \item{y}{ (subtracting \code{offset} if necessary), }
  \item{parms}{ as input, }
  \item{numresp}{ the number of responses, here \code{1}, }
  \item{summary}{ a function to be invoked by \code{summary.rpart}, }
  \item{text}{ a function to be invoked by \code{text.rpart}. }
}
}
\description{
Set up methods for an \code{rpart} model.
}
\details{
These functions are to be used only by the function \code{rpart}.
}
\section{Functions}{
\itemize{
\item \code{rpart.anova()}: 

\item \code{rpart.mrt()}: 

\item \code{rpart.poisson()}: 

\item \code{rpart.class()}: 

\item \code{rpart.dist()}: 

\item \code{rpart.exp()}: 

}}
\keyword{internal}
