% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.pca.R
\name{rpart.pca}
\alias{rpart.pca}
\title{Principle Components Plot of a Multivariate Rpart Object}
\usage{
rpart.pca(
  tree,
  pts = TRUE,
  plt.allx = TRUE,
  speclabs = TRUE,
  specvecs = TRUE,
  wgt.ave = FALSE,
  add.tree = TRUE,
  cv1 = 1L,
  cv2 = 2L,
  chulls = TRUE,
  interact = FALSE,
  ...
)
}
\arguments{
\item{tree}{A fitted object of class \code{rpart} containing a multivariate
regression tree.}

\item{pts}{If \code{TRUE}, large points representing the leaf means are
plotted.}

\item{plt.allx}{If \code{TRUE}, small points representing individual cases
are plotted.}

\item{speclabs}{If \code{TRUE} the labels of the response variables are
plotted.}

\item{specvecs}{If \code{TRUE} the vectors of the response variables are
plotted provided \code{wgt.ave} is \code{FALSE}.}

\item{wgt.ave}{If \code{TRUE} use weighted averages of responses not vectors.}

\item{add.tree}{If \code{TRUE} add the tree structure to the plot.}

\item{cv1}{Defines the principal component to plot horizontally -- but see
\code{interact}.}

\item{cv2}{Defines the principal component to plot vertically -- but see
\code{interact}.}

\item{chulls}{If \code{TRUE} adds convex hulls to thr tree groups.}

\item{interact}{If \code{TRUE} the plot can be viewed in dimensions by
left-clicking to top-left, bottom-right or bottom-left (reset).}

\item{...}{Arguments to be passed to or from other functions or methods.}
}
\value{
\code{NULL} (well, this is not really true).
}
\description{
Plots a PCA of the rpart object on the current graphics device.
}
\details{
This function plots a PCA biplot of the group means (leaves) of
multivariate regresssion objects of class \code{rpart}. The responses and
group means and indivdual cases can be shown on the plot. If responses are
positive (e.g., species-environment data) weighted averages of responses can
be plotted.

The PCA biplot plot is produced on the current graphics device.
}
\examples{
data(spider)
fit <- mvpart(data.matrix(spider[,1:12])~herbs+reft+moss+sand+twigs+water,spider)
rpart.pca(fit)
rpart.pca(fit, wgt.ave=TRUE, interact=TRUE)

}
\seealso{
\code{\link{rpart}}
}
\keyword{tree}
