% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snip.rpart.R
\name{snip.rpart}
\alias{snip.rpart}
\title{Snip Subtrees of an Rpart Object}
\usage{
snip.rpart(x, toss)
}
\arguments{
\item{x}{An object of class \code{rpart}. This is assumed to be the result
of some function that produces an object with the same named components as
that returned by the \code{rpart} function.}

\item{toss}{An integer vector containing indices (node numbers) of all
subtrees to be snipped off.  If missing, user selects branches to snip off as
described below.}
}
\value{
A \code{rpart} object containing the nodes that remain after
specified or selected subtrees have been snipped off.
}
\description{
Creates a "snipped" rpart object, containing the nodes that remain after
selected subtrees have been snipped off.  The user can snip nodes using the
toss arguement, or interactively by clicking the mouse button on specified
nodes within the graphics window.
}
\details{
A dendrogram of \code{rpart} is expected to be visible on the
graphics device, and a graphics input device (e.g., a mouse) is required.
Clicking (the selection button) on a node displays the node number, sample
size, response yvalue, and Error (dev). Clicking a second time on the same
node snips that subtree off and visually erases the subtree. This process may
be repeated an number of times.  Warnings result from selecting the root or
leaf nodes.  Clicking the exit button will stop the snipping process and
return the resulting \code{rpart} object.

Warning: visually erasing the plot is done by over-plotting with the
background colour.  This will do nothing if the background is transparent
(often true for screen devices).

See the documentation for the specific graphics device for details on
graphical input techniques.
}
\seealso{
\code{plot.rpart}
}
\keyword{tree}
