% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xdiss.R
\name{xdiss}
\alias{xdiss}
\title{Extendend Dissimilarity Measures}
\usage{
xdiss(
  data,
  dcrit = 1,
  dauto = TRUE,
  dinf = 0.5,
  method = "man",
  use.min = TRUE,
  eps = 1e-04,
  replace.neg = TRUE,
  big = 10000,
  sumry = TRUE,
  full = FALSE,
  sq = FALSE
)
}
\arguments{
\item{data}{Data matrix.}

\item{dcrit}{Dissimilarities < \code{dcrit} are considered to have no species
in common and are recalculated.}

\item{dauto}{Automatically select tuning parameters -- recommended.}

\item{dinf}{Internal parameter -- leave as is usually.}

\item{method}{Dissimilarity index.}

\item{use.min}{Minimum dissimilarity of pairs of distances used --
recommended.}

\item{eps}{Internal parameter -- leave as is usually.}

\item{replace.neg}{Internal parameter -- leave as is usually.}

\item{big}{Internal parameter -- leave as is usually.}

\item{sumry}{Whether to print summary of extended dissimilarities.}

\item{full}{Return the square dissimilarity matrix.}

\item{sq}{Square the dissimilarities -- useful for distance-based
partitioning.}
}
\value{
Returns an object of class distance with attributes "Size" and "ok".
"ok" is TRUE if rows are not disconnected (De'ath 1999).
}
\description{
The function computes extended dissimilarity indices which are for long
gradients have better good rank-order relation with gradient separation and
are thus efficient in community ordination with multidimensional scaling.
}
\details{
The function knows the same dissimilarity indices as
\code{\link{gdist}}.
}
\examples{
data(spider)
spider.dist <- xdiss(spider[,1L:12L])

}
\references{
De'ath, G. (1999) Extended dissimilarity: a method of robust estimation of
ecological distances from high beta diversity data. Plant Ecology 144(2):
191-199.

Faith, D.P, Minchin, P.R. and Belbin, L. (1987) Compositional dissimilarity
as a robust measure of ecological distance. Vegetatio 69: 57-68.
}
\author{
Glenn De'ath
}
\keyword{multivariate}
