% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpred.rpart.R
\name{xpred.rpart}
\alias{xpred.rpart}
\title{Return Cross-Validated Predictions}
\usage{
xpred.rpart(fit, xval = 10, cp)
}
\arguments{
\item{fit}{An \code{rpart} object.}

\item{xval}{Number of cross-validation groups. This may also be an explicit
list of integers that define the cross-validation groups.}

\item{cp}{The desired list of complexity values. By default it is taken from
the \code{cptable} component of the fit.}
}
\value{
A matrix with one row for each observation and one column for each
complexity value.
}
\description{
Gives the predicted values for an \code{rpart} fit, under cross validation,
for a set of complexity parameter values.
}
\details{
Complexity penalties are actually ranges, not values. If the \code{cp} values
found in the table were .36, .28, and .13, for instance, this means that the
first row of the table holds for all complexity penalties in the >=.36, <=1
range, the second row for \code{cp} in the >=.28, <.36 and the third row for
>=.13 <.28. By default, the geometric mean of each interval is used for cross
validation.
}
\examples{
data(car.test.frame)
fit <- rpart(Mileage ~ Weight, car.test.frame)
xmat <- xpred.rpart(fit)
xerr <- (xmat - car.test.frame$Mileage)^2
apply(xerr, 2, sum)   # cross-validated error estimate

# approx same result as rel. error from printcp(fit)
apply(xerr, 2, sum)/var(car.test.frame$Mileage)
printcp(fit)

}
\seealso{
\code{\link{rpart}}
}
\keyword{tree}
